Alfred

+++ controller/mealController.js
@@ -18,45 +18,32 @@ exports.CreateMeal = async (req, res) => {
 
     image = req.file;
 
+    // console.log(req.body, image);
     const restaurant = await RestaurantModel.findById(restaurantId);
-
     if (!restaurant) {
       if (image && image.path) {
-        fs.unlinkSync(image.path);
+        if (fs.existsSync(image.path)) fs.unlinkSync(image.path);
       }
       return res.status(404).json({
         message: "Restaurant not found, meal must belong to a restaurant",
       });
     }
     let mealImage = null;
-
-    switch (true) {
-      case !name:
-        return res.status(400).json({
-          message: "Please provide a name for the meal",
-        });
-      case !image:
-        return res.status(400).json({
-          message: "Please provide an image for the meal",
-        });
-      case !description:
-        return res.status(400).json({
-          message: "Please provide a description for the meal",
-        });
-      case !price:
-        return res.status(400).json({
-          message: "Please provide a price for the meal",
-        });
-      default:
-        if (image && image.path) {
-          mealImage = await uploadCloudinaryImage(image.path);
-          fs.unlinkSync(image.path);
-        }
+    if (!name || !image || !description || !price) {
+      if (image && image.path) {
+        mealImage = await uploadCloudinaryImage(image.path);
+        if (fs.existsSync(image.path)) fs.unlinkSync(image.path);
+        image = null;
+      }
+      return res.status(400).json({
+        message:
+          "Please provide all the required fields (name, image, description, price)",
+      });
     }
 
     if (image && image.path) {
       mealImage = await uploadCloudinaryImage(image.path);
-      fs.unlinkSync(image.path);
+      if (fs.existsSync(image.path)) fs.unlinkSync(image.path);
     }
     const meal = new MealModel({
       name,
@@ -72,7 +59,7 @@ exports.CreateMeal = async (req, res) => {
     });
   } catch (error) {
     if (image && image.path) {
-      fs.unlinkSync(image.path);
+      if (fs.existsSync(image.path)) fs.unlinkSync(image.path);
     }
     console.log(error);
     res.status(500).json({
@@ -81,7 +68,8 @@ exports.CreateMeal = async (req, res) => {
     });
   }
 };
-exports.getAll = async (req, res) => {
+
+exports.getAllMeals = async (req, res) => {
   try {
     const meal = await MealModel.find();
 
@@ -99,7 +87,7 @@ exports.getAll = async (req, res) => {
     });
   }
 };
-exports.getOne = async (req, res) => {
+exports.getOneMeal = async (req, res) => {
   try {
     const { id } = req.params;
     const meal = await MealModel.findById(id);
@@ -120,7 +108,7 @@ exports.getOne = async (req, res) => {
     });
   }
 };
-exports.update = async (req, res) => {
+exports.updateMeal = async (req, res) => {
   let image = null;
   try {
     const { id, restaurantId } = req.params;
@@ -136,7 +124,7 @@ exports.update = async (req, res) => {
 
     if (!restaurant) {
       if (image && image.path) {
-        fs.unlinkSync(image.path);
+        if (fs.existsSync(image.path)) fs.unlinkSync(image.path);
       }
       return res.status(404).json({
         message: "Restaurant not found, meal must belong to a restaurant",
@@ -148,24 +136,29 @@ exports.update = async (req, res) => {
     const meal = await MealModel.findById(id);
     if (!meal) {
       if (image && image.path) {
-        fs.unlinkSync(image.path);
+        if (fs.existsSync(image.path)) fs.unlinkSync(image.path);
       }
       return res.status(404).json({
         message: "meal not found",
       });
     }
+
     if (image && image.path) {
       mealImage = await uploadCloudinaryImage(image.path);
-      fs.unlinkSync(image.path);
+      if (fs.existsSync(image.path)) fs.unlinkSync(image.path);
+      // image = null;
     }
     Object.assign(meal, { name, image: mealImage, description, price });
-    await meal.save();
+    const updatedMeal = await MealModel.findByIdAndUpdate(id, meal, {
+      new: true,
+    });
     res.status(200).json({
       message: "Meal updated successfully",
+      data: updatedMeal,
     });
   } catch (error) {
     if (image && image.path) {
-      fs.unlinkSync(image.path);
+      if (fs.existsSync(image.path)) fs.unlinkSync(image.path);
     }
     console.log(error);
     res.status(500).json({
@@ -174,7 +167,7 @@ exports.update = async (req, res) => {
     });
   }
 };
-exports.delete = async (req, res) => {
+exports.deleteMeal = async (req, res) => {
   try {
     const { id } = req.params;
     const meal = await MealModel.findByIdAndDelete(id, { new: true });
@@ -201,6 +194,7 @@ exports.delete = async (req, res) => {
 
 exports.getRestaurantMeals = async (req, res) => {
   try {
+    // console.log("I am working!");
     const { restaurantId } = req.params;
     const restaurant = await RestaurantModel.findById(restaurantId);
     if (!restaurant) {
@@ -208,7 +202,7 @@ exports.getRestaurantMeals = async (req, res) => {
         message: "Restaurant not found",
       });
     }
-    const meal = await MealModel.findOne({ restaurantId });
+    const meal = await MealModel.find({ restaurantId });
     const total = meal.length;
     res.status(200).json({
       message:
